requirements.txt
dask==0.11.0
decorator==4.0.10
flake8==3.3.0
isort==4.2.5
networkx==1.11
numpy==1.11.1
matplotlib==1.5.3
Pillow==3.3.1
PyYAML==3.12
six==1.10.0
pytest==3.0.4
toolz==0.8.0
git+https://github.com/fchollet/keras.git@73bf06fb023a8b37ddf2e2a168bbf920c7a6c766
h5py==2.6.0
git+https://github.com/nlhepler/pydot.git
tox==2.4.1
tensorflow==1.0.0
